fast-model-main/requirements.txt
alembic==1.8.1
antlr4-python3-runtime==4.9.3
black==22.8.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
databricks-cli==0.17.3
docker==5.0.3
entrypoints==0.4
flake8==5.0.4
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
mccabe==0.7.0
mlflow==1.28.0
mypy-extensions==0.4.3
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.2.3
packaging==21.3
pandas==1.4.4
pathspec==0.10.1
platformdirs==2.5.2
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
protobuf==4.21.5
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.1
scipy==1.9.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.40
sqlparse==0.4.2
tabulate==0.8.10
tomli==2.0.1
typer==0.6.1
typing_extensions==4.3.0
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
zipp==3.8.1
